tidy up file system
