+++ old/index.js
@@ -1,42 +1,27 @@
 /* eslint-disable no-mixed-spaces-and-tabs */
 // Require the necessary discord.js classes
-const { Client, Collection, Events, GatewayIntentBits, Partials } = require('discord.js');
+const fs = require('node:fs');
+const path = require('node:path');
+const client = require('../handlers/newClient');
+const { Client, Events, GatewayIntentBits, Partials, MessageActionRow, MessageButton } = require('discord.js');
 const { OpenAI } = require('openai');
 require('dotenv/config');
-const fs = require('fs');
-
 const token = process.env.TOKEN;
 
 // Create a new client instance
-const client = new Client({
-	intents: [
-		GatewayIntentBits.Guilds,
-		GatewayIntentBits.GuildMembers,
-		GatewayIntentBits.GuildMessages,
-		GatewayIntentBits.MessageContent,
-		GatewayIntentBits.DirectMessages,
-	],
-	partials: [
-		Partials.Channel,
-		Partials.Message,
-	],
-});
-
-
-
-client.commands = new Collection();
-
-//command handler
-//retrieves the command files in an array
-const commandFiles = fs.readdirSync('./commands').filter(file => file.endsWith('.js'));
-
-for (const file of commandFiles) {
-	const command = require(`./commands/${file}`);
-	// set a new item in the Collection
-	// with the key as the command name and the value as the exported module
-	client.commands.set(command.data.name, command);
-}
-
+// const client = new Client({
+// 	intents: [
+// 		GatewayIntentBits.Guilds,
+// 		GatewayIntentBits.GuildMembers,
+// 		GatewayIntentBits.GuildMessages,
+// 		GatewayIntentBits.MessageContent,
+// 		GatewayIntentBits.DirectMessages,
+// 	],
+// 	partials: [
+// 		Partials.Channel,
+// 		Partials.Message,
+// 	],
+// });
 
 // When the client is ready, run this code (only once).
 // The distinction between `client: Client<boolean>` and `readyClient: Client<true>` is important for TypeScript developers.
@@ -44,17 +29,7 @@ for (const file of commandFiles) {
 client.once(Events.ClientReady, readyClient => {
 	console.log(`Ready logged! Logged in as ${readyClient.user.tag}`);
 
-    command(client, 'createtextchannel', (message) => {
-		const name = message.content.replace('!createtextchannel ', '')
-
-		message.guild.channels
-			.create(name, {
-				type: 'text',
-			})
-			.then((channel) => {
-				console.log(channel)
-			})
-	})
+	
 });
 
 // Use API Key Directly
@@ -63,7 +38,9 @@ const openai = new OpenAI({ apiKey: process.env.API_KEY });
 // Event listener for incoming messages
 client.on('messageCreate', async msg => {
 	// console.log({msg});
-
+	
+	
+    
 	// Ignore messages from bots and empty messages
 	if (msg.author.bot || !msg.content) return;
 
@@ -77,8 +54,8 @@ client.on('messageCreate', async msg => {
 
 	// Show the bot typing as we wait for a response
 	await msg.channel.sendTyping();
-
-	console.log(msg);
+	
+	//console.log(msg);
 
 	// Empty array to contain whole conversation, so bot can refer back
 	const conversation = [];
@@ -87,10 +64,12 @@ client.on('messageCreate', async msg => {
 	try {
 		const conversationHistory = await msg.channel.messages.fetch({ limit: 10 });
 		conversationHistory.reverse();
-
+		// console.log("Client:", client);
+		
 		// For each message fetched, it checks who sent it and pushes to the conversation array
 		conversationHistory.forEach((message) => {
 			if (message.author.bot && message.author.id !== client.user.id) return;
+			//console.log("Client:", client.user);
 
 			// If the author is our bot
 			if (message.author.id === client.user.id) {
@@ -130,7 +109,7 @@ client.on('messageCreate', async msg => {
 			// Direct message the user
 			try {
 				const user = await client.users.fetch(msg.author.id);
-				console.log(`userName: ${user}`);
+				//console.log(`userName: ${user}`);
 				user.send(`In response to your message: "${msg.content}", the AI says: "${response}"`)
 					//.then(msg => console.log(`Sent message: ${msg.content} to ${user}`))
 					.catch(console.error);
@@ -150,17 +129,41 @@ client.on('messageCreate', async msg => {
   	}
 });
 
-// Event listener for interactions
-client.on(Events.InteractionCreate, async interaction => {
-    if (!interaction.isButton() && interaction.componentType !== 'SELECT_MENU') return;
 
-    if (interaction.isButton() && interaction.customId === 'primaryButton') {
-        await interaction.reply('Primary button clicked!');
-    } else if (interaction.componentType === 'SELECT_MENU' && interaction.customId === 'selectMenu') {
-        await interaction.reply(`Selected option: ${interaction.values.join(', ')}`);
+// load the events files on startup
+const eventsPath = path.join(__dirname, "events");
+const eventFiles = fs
+    .readdirSync(eventsPath)
+    .filter((file) => file.endsWith(".js"));
+
+for (const file of eventFiles) {
+    const filePath = path.join(eventsPath, file);
+    const event = require(filePath);
+    if(event.once) {
+        client.once(event.name, (...args) => event.execute(...args));
+    } else {
+        client.on(event.name, (...args) => event.execute(...args));
     }
-});
+}
+
+// load the command files on startup
+// client.commands = new Collection();
+// const commandsPath = path.join(__dirname, "commands");
+// const commandFiles = fs
+//     .readdirSync(commandsPath)
+//     .filter((file) => file.endsWith(".js"));
+
+// for (const file of commandFiles) {
+//     const filePath = path.join(commandsPath, file);
+//     const command = require(filePath);
+//     if("data" in command && "execute" in command) {
+//         client.commands.set(command.data.name, command); 
+//     } else {
+//         console.log(
+//             `[WARNING] The command at ${filePath} is missing a required "data" or "execute property"`
+//         );
+//     }
+// }
 
 // Log in to Discord with your client's token
 client.login(token);
-
